Arreglo: Al procesar el financiero y si arroja error igual se graba financiero y movimientos y quedan saldos inconsistentes, se elimina el commit. Arreglo: Al generar saldos vencidos, no se sumaba el monto de cargos de cobranza al monto total. Arreglo: Al generar saldos vencidos con fecha valor, se volvía a generar los saldos con la fecha actual y no con la fecha valor. Arreglo: Error al consultar si una categoria permite saldo negativo o no. Arreglo: Al ejecutar el proceso de respaldo, se grabar la fecha contable mostrada en el campo, pudiendo ser esta una fecha diferente. Nuevo: Nuevo campo en AccountOverdueBalanace, se agrega el campo fecha de pago de la cuota (payment_date) Nuevo: Nuevo proceso en lote, depositos en banco por recaudacion. Nuevo: Nueva validación al pagar una cuota, si la cuota tiene fecha de pago, lanza una error de ejecucion. Nuevo: Luego de efectuar el pago de cuota, se valida el saldo consolidado total de la cuota es cero y se la pone como pagada, y ya no solo se compara el saldo de CAPITAL. Mejora: Al generar saldos vencidos, en cuotas proyectadas y por vencer se registra las diferentes fechas de pagos de la tabla de pagos. Mejora: Se optimiza las consultas de los detalles solicitados, con error, satisfactorios y totales de un proceso batch. Mejora: En listados de saldos vencidos se muestra la fecha de pago de la cuota. Mejora: Se resetea el formulario en la acción de cierre de día, con el fin de volver al modo lista y que se actualice la fecha contable.

+++ powerfin/src/com/powerfin/model/BatchProcessType.java
@@ -147,7 +147,7 @@ public class BatchProcessType implements Serializable {
 		}
 	}
 	
-	public Integer getCountErrorDetails() throws Exception
+	public Long getCountErrorDetails() throws Exception
 	{
 		try
 		{
@@ -166,7 +166,7 @@ public class BatchProcessType implements Serializable {
 		}
 	}
 	
-	public Integer getCountSatisfactoryDetails() throws Exception
+	public Long getCountSatisfactoryDetails() throws Exception
 	{
 		try
 		{
@@ -185,7 +185,7 @@ public class BatchProcessType implements Serializable {
 		}
 	}
 	
-	public Integer getCountRequestDetails() throws Exception
+	public Long getCountRequestDetails() throws Exception
 	{
 		try
 		{
@@ -204,7 +204,7 @@ public class BatchProcessType implements Serializable {
 		}
 	}
 	
-	public Integer getCountTotalDetails() throws Exception
+	public Long getCountTotalDetails() throws Exception
 	{
 		try
 		{
